refactor(tests): lower linecheck gate to 2500 lines; split service_tests.rs again
Lowers the pre-push linecheck limit from 3000 → 2500.  service_tests.rs
(2677 lines) still exceeded the new ceiling, so tags/machines/model tests
are moved to service_model_tests.rs, bringing service_tests.rs to 2467 lines.
All 760 tests remain green.
