+++ .githooks/pre-push
@@ -86,7 +86,7 @@ fi
 
 echo "Checking .rs files do not exceed 3000 lines..."
 # shellcheck disable=SC2046
-linecheck --max-lines 3000 $(find src -name '*.rs')
+linecheck --max-lines 2500 $(find src -name '*.rs')
 echo "Line-count check passed."
 
 if [ "${SKIP_CHANGELOG:-0}" != "1" ]; then
